Modify line preprocessing.
In particular, modify generation of \t() from \fad(). Now it takes into
account \alpha values in the line (and \1a from the style, if there
aren't any in the line).  The preprocessing is a relic of an ancient
version of the script archictecture and hasn't been phased out
completely because I'm too lazy to expend the effort to rewrite it
properly (that, and if I did that I'd feel bad about not rewriting all
of the other stuff that needs it).

Also generally give functions more verbose names and reorganize them
pointlessly.