Misc. code cleanup (by tophf)
Disambiguate multiple assignment by a function with implicit
parentheses to improve readability rather than code correctness (e.g.
`x, y = func a, b` becomes `x, y = func(a, b)`

Fix a few instances of `a-b` which compiles to `a(-b)` because
moonscript is batshit insane (this has actually been fixed in
moonscript master (commit 4b5d6f1) but is not yet in an actual
versioned version).

Move the frame-by-frame function next to the other application loop
functions to more logically structure the code.