Fix comparison to nil.
More batshit insanity from moonscript here.

To do forward declarations, you have to prefix the variable identifier
with `local`. But `local` is also used for shadowing, but it also has
to be a forward declaration if you are shadowing a variable. Prefixing
assignment with `local` does not work.

And if you forget to put it before your forward declarations, well,
then it attempts to do something like `local _ = fdeclvar`. What.