Fix a bug with trim GUI calculating frame length.
Apparently this has been around forever and I didn't know because
nobody pointed it out to me and I don't use the Trim gui because it's
for retards on 2.1.9 which isn't even supported any more so I don't
know why I haven't deleted it entirely.

A cookie to rcombs for pointing this out to me.