Fix independent clip tracking bug introduced in 7404111f92.
This has been around for a while now. Apparently this feature doesn't get
used too often. The problem was that the main dialog config variable was
being overwritten by the clip dialog window and deleting the tracking data.
The solution was to store the dialog configurations as separate members of a
table.

Thanks to __ar for pointing this bug out to me.