Add clipOnly option.
Allows clips to be tracked from the main dialog without moving the line.

The same behavior can be had by using the separate clip tracking
feature and either disabling all of the options in the main window or
not providing tracking data to the main window. However, this may be
more convenient for the user.