Trimhandler: fix unicode on windows for real.
Big thanks to @line0 and @tophf for suggestions and information about
how to fix this.

Extra thanks to @line0 for putting up with me being a vitriol-spewing,
obnoxious manbaby while repeatedly failing to implement it properly.

That said, I'm never going to ever touch this again. It works for me,
and if it doesn't work for you, patches welcome.

Fixes #16.