MotionHandler: crush all times down to centiseconds.
Apparently a chunk of the error in transforms was being caused by not
paying attention to how Aegisub handles times.

Which is odd, because I thought this was working correctly when I
tested it previously. Either way, this produces accurate results with
Aegisub 3.2.2. I expect previous tests were not rigorous enough.