Aegisub-Motion: tweak accelerators a bit.
Was done in the hope of fixing #24, which it did... on OS X.
Unfortunately, ctrl-a to select all in the text box still does not work
on Windows, at least for me, so I can only assume it's somehow Aegisub's
fault. Since it fixes the issue on at least one platform and doesn't
cause different behavior on the main one, might as well commit the
change.