TrimHandler: misc. Windows fixes.
$? does not work here because for some reason powershell always throws
an exception with the first line of the encoder output. This exception
gets mysteriously steamrolled and doesn't actually abort the encoding
process, but it does set error flags. Fortunately $LASTEXITCODE seems
to work okay.

The log file needs to be manually cleared out because it will
simply be appended to between runs and will contain the same message
multiple times if consecutive runs cause the mystery error condition.

Also apparently the log file was leaking a file handle. Whoops.