Improve fade handling by blending into alphas.
ASS allows \fad(e) and \t(\alpha) tags to coexist in a single line, and
in such a case, the opacities are multiplied appropriately. Aegisub-
Motion previously would break lines using this "feature" by converting
\fad tags to transforms automagically, in which case the opacities would
not be blended.

This fixes the problem by manually multiplying out the opacities using a
piecewise trapezoidal function to calculate the current fade state. It
also tries to ensure the necessary alpha tags are present in the line
before processing.